requirements.txt
pydantic-ai
faiss-cpu
sentence-transformers
PyMuPDF
google-generativeai
